splitwise/backend/requirements.txt
Flask==2.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-CORS==3.0.10
Flask-Bcrypt==0.7.1
Flask-Migrate==3.1.0
Werkzeug==2.0.1
email-validator==1.3.1
Pillow==9.5.0
psycopg2-binary==2.9.5
SQLAlchemy==1.4.0
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
MarkupSafe==3.0.2
typing-extensions==4.13.2